finish dividing UI in components

+++ client/src/components/TodoList.js
@@ -6,6 +6,12 @@ const API_BASE = "http://localhost:5000/api/v1";
 
 function TodoList() {
   const [todos, setTodos] = useState([]);
+  
+  const [todoToEdit, setTodoToEdit] = useState({
+    id: null,
+    description: "",
+    duedate: ""
+  });
 
   useEffect(() => {
     getTodos();
@@ -33,6 +39,21 @@ function TodoList() {
     setTodos([...todos, data]);
   }
 
+  const updateTodo = async (id, description, duedate) => {
+    await fetch(API_BASE + "/todos/update/" + id, {
+      method: 'PATCH',
+      headers: {
+        "Content-Type": "application/json"
+      },
+      body: JSON.stringify({
+        description: description,
+        duedate: duedate
+      })
+    }).then(res => res.json());
+
+    getTodos();
+  }
+
   const completeTodo = async id => {
     const data = await fetch(API_BASE + "/todos/update/done/" + id, {
       method: 'PATCH'
@@ -48,21 +69,36 @@ function TodoList() {
   }
 
   const deleteTodo = async id => {
-    const data = await fetch(API_BASE + "/todos/delete/" + id, {
-      method: 'DELETE'
-    }).then(res => res.json());
+    const canDelete = window.confirm("Deseja Excluir Esta Tarefa?");
 
-    setTodos(todos => todos.filter(todo => todo._id !== data._id));
+    if (canDelete) {
+      const data = await fetch(API_BASE + "/todos/delete/" + id, {
+        method: 'DELETE'
+      }).then(res => res.json());      
+
+      setTodos(todos => todos.filter(todo => todo._id !== data._id));
+    }
+  }
+
+  if (todoToEdit.id) {
+    return (
+      <TodoForm
+        todoToEdit={ todoToEdit }
+        setTodoToEdit={ setTodoToEdit }
+        onSubmit={ updateTodo }
+      />
+    );
   }
 
   return (
     <div>
-      <TodoForm onSubmit={ createTodo }/>
+      <TodoForm todoToEdit={ todoToEdit } onSubmit={ createTodo }/>
       {todos.map((todo, index) => (
         <Todo 
           todo={ todo }
           completeTodo={ completeTodo }
           deleteTodo={ deleteTodo }
+          setTodoToEdit = { setTodoToEdit }
           key={ index }
         />
       ))}
